--- Auto-Git Commit ---

+++ backend/controllers/sauce.js
@@ -68,16 +68,16 @@ exports.likedNoliked = (req, res, next) => {
 exports.deleteSauce = (req, res, next) => {
   Sauce.findOne({ _id: req.params.id })
     .then(sauce => {
-    //  if (sauce.userId != req.auth.userId) {
-    //    res.status(401).json({ message: 'Not authorized' });
-    //  } else {
+      if (sauce.userId != req.auth.userId) {
+        res.status(401).json({ message: 'Not authorized' });
+      } else {
         const filename = sauce.imageUrl.split('/images/')[1];
         fs.unlink(`images/${filename}`, () => {
           Sauce.deleteOne({ _id: req.params.id })
             .then(() => res.status(200).json({ message: 'Sauce supprimée !' }))
             .catch(error => res.status(401).json({ error }));
         });
-    //  }
+      }
     })
     .catch(error => {
       res.status(500).json({ error });
